# TodoApi/TodoApi/Models/Game.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace TodoApi.Model
{
    public class Game
    {
        public int Id_game { get; set; }
        public string Game_name { get; set; }
    }
}