# clases.js
class User{
    name;
    surname;
    username;
    email;
    password;
    id_city;
    profile;

    constructor(Name, Surname, Username, Email, Password, Id_city, Profile) {
        this.name = Name;
        this.surname = Surname;
        this.username = Username;
        this.email = Email;
        this.password = Password;
        this.id_city = Id_city;
        this.profile = Profile;
    }

    
}

class UserId{
    id_user;
    name;
    surname;
    username;
    email;
    password;
    id_city;
    profile;

    constructor(Id_user, Name, Surname, Username, Email, Password, Id_city, Profile) {
        this.id_user = Id_user;
        this.name = Name;
        this.surname = Surname;
        this.username = Username;
        this.email = Email;
        this.password = Password;
        this.id_city = Id_city;
        this.profile = Profile;
    }
}